% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensd.R
\name{gensd}
\alias{gensd}
\title{Attributes skewed and dependent}
\usage{
gensd (size, sk, dep, ...)
}
\arguments{
\item{size}{a vector of two elements specifying respectively the number of rows and columns of the matrix of attributes to be generated.}

\item{sk}{a scalar that represents the skewness coefficient to be considered.}

\item{dep}{a scalar that represents the dependence correlation between values within a given attribute.}
}
\value{
Returns a list:
\item{spadata}{a matrix containing expected data.}
\item{init}{a recall of initial parameters}
}
\description{
Generates spatial attributes that are skewed and dependent.
}
\examples{

}
\author{
Sewanou Honfo \url{<honfosewanou@gmail.com>}  and Brezesky Kotanmi \url{<kotangaebrezesky@gmail.com>} and Eunice Gnanvi \url{<eunysgnanvi@gmail.com>} / LABEF_08_2019
}
