% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensd.R
\name{gensd}
\alias{gensd}
\title{Spatial dependent attribute with fixed skewness}
\usage{
gensd (size, grid, sk = 1, dep = 0.5, mu = 0, v = 1,  method = "SAR", dst = "skewnormal")
}
\arguments{
\item{size}{a vector of two elements specifying respectively numbers of indivuals and time points of the attribute to be generated.}

\item{grid}{number of rows and columns for grids.}

\item{sk}{a scalar that represents the skewness coefficient to be considered.}

\item{dep}{a scalar that represents the dependence correlation between values within a given attribute.}

\item{mu}{the mean value}

\item{v}{the variance value}

\item{method}{a character string naming the the approach used to create spatial dependence between variables. The "SAR" method is the Kapoor, Kelejian, and Prucha (2007)-spatial autoregressive random effects (SAR-RE) model. The "SARMA" method represents the Lee and Yu (2012) spatial autoregressive moving average random effects (SARMA-RE) model.}

\item{dst}{a character string naming the distribution to be used for vectors generation. "skewnormal" the skew-normal distribution is limited to a range of [0,1[ for the coefficient of skewness. "lognormal" the lognormal distribution allows a large range of coefficients skewness values.}
}
\value{
Returns a gsd object.
}
\description{
Generates a spatial dependent attribute with a fixed skewness.
}
\author{
Sewanou Honfo \url{<honfosewanou@gmail.com>}  and Brezesky Kotanmi \url{<kotangaebrezesky@gmail.com>} and Eunice Gnanvi \url{<eunysgnanvi@gmail.com>} and Chenangnon Tovissode \url{<chenangnon@gmail.com>} and Romain glèlè Kakaï \url{<glele.romain@gmail.com>}   / LABEF_08_2019
}
